Lock down fbjs version to prevent potential sad times in the future.
From the fbjs description:

> If you are consuming the code here and you are not also a Facebook project,
be prepared for a bad time. APIs may appear or disappear and we may not
follow semver strictly, though we will do our best to. This library is
being published with our use cases in mind and is not necessarily meant to
be consumed by the broader public.

Seems like reason enough to not depend on 0.8.x over 0.8.0.

https://github.com/facebook/fbjs#purpose